- moved weights_initializations.py to utility.py - changed unit_test name from test_weights_initializations.py to test_utility.py
